Refactored models to reduce circular import mitigations
