+++ examples/custom-scripts/NetBox_DNS_Tenancy.py
@@ -3,7 +3,8 @@
 from extras.scripts import Script
 from ipam.models import IPAddress
 
-from netbox_dns.models import Record, RecordTypeChoices
+from netbox_dns.models import Record
+from netbox_dns.choices import RecordTypeChoices
 
 name = "NetBox DNS Tenancy Synchronizer"
