+++ netbox_dns/choices/__init__.py
@@ -1,0 +1,2 @@
+from .record import *
+from .zone import *
