+++ netbox_dns/choices/zone.py
@@ -1,0 +1,17 @@
+from utilities.choices import ChoiceSet
+
+
+class ZoneStatusChoices(ChoiceSet):
+    key = "Zone.status"
+
+    STATUS_ACTIVE = "active"
+    STATUS_RESERVED = "reserved"
+    STATUS_DEPRECATED = "deprecated"
+    STATUS_PARKED = "parked"
+
+    CHOICES = [
+        (STATUS_ACTIVE, "Active", "blue"),
+        (STATUS_RESERVED, "Reserved", "cyan"),
+        (STATUS_DEPRECATED, "Deprecated", "red"),
+        (STATUS_PARKED, "Parked", "gray"),
+    ]
