+++ netbox_dns/template_content.py
@@ -5,7 +5,8 @@ from django.conf import settings
 from netbox.plugins.utils import get_plugin_config
 from netbox.plugins import PluginTemplateExtension
 
-from netbox_dns.models import Record, RecordTypeChoices, Zone, View, NameServer
+from netbox_dns.models import Record, Zone, View, NameServer
+from netbox_dns.choices import RecordTypeChoices
 from netbox_dns.tables import RelatedRecordTable
