+++ netbox_dns/tests/record/test_api.py
@@ -5,7 +5,8 @@ from rest_framework import status
 from utilities.testing import APIViewTestCases
 
 from netbox_dns.tests.custom import APITestCase, NetBoxDNSGraphQLMixin
-from netbox_dns.models import View, Zone, NameServer, Record, RecordTypeChoices
+from netbox_dns.models import View, Zone, NameServer, Record
+from netbox_dns.choices import RecordTypeChoices
 
 
 class RecordAPITestCase(
