+++ netbox_dns/tests/record/test_name_validation.py
@@ -1,7 +1,8 @@
 from django.test import TestCase, override_settings
 from django.core.exceptions import ValidationError
 
-from netbox_dns.models import NameServer, Zone, Record, RecordTypeChoices
+from netbox_dns.models import NameServer, Zone, Record
+from netbox_dns.choices import RecordTypeChoices
 
 
 class RecordNameValidationTestCase(TestCase):
