+++ netbox_dns/tests/zone/test_auto_soa.py
@@ -2,13 +2,8 @@ from dns import rdata
 
 from django.test import TestCase
 
-from netbox_dns.models import (
-    NameServer,
-    Record,
-    RecordClassChoices,
-    RecordTypeChoices,
-    Zone,
-)
+from netbox_dns.models import NameServer, Record, Zone
+from netbox_dns.choices import RecordClassChoices, RecordTypeChoices
 
 
 def parse_soa_value(soa):
