+++ netbox_dns/tests/zone/test_views.py
@@ -1,7 +1,8 @@
 from utilities.testing import ViewTestCases, create_tags
 
 from netbox_dns.tests.custom import ModelViewTestCase
-from netbox_dns.models import NameServer, View, Zone, ZoneStatusChoices
+from netbox_dns.models import NameServer, View, Zone
+from netbox_dns.choices import ZoneStatusChoices
 
 
 class ZoneViewTestCase(
